Remove Multimethod library
Not that I don't love it, but these are closed algorithms, so I
don't really need it.  Gone is the need to make things lovely
and pretty to such a strange extent; it's better to get rid of
dependencies, pending a move to make this a viable browser-side
module.

Also, dispatching by indexOf instead of dispatching against
a string key is intuitively not the best move.

fixes #11